test: send bitcoin confirmation storie mocks moved to tests files
